Add Legion War event as a SpecialEvent type with KaiserBattle (with a help of ShizuruNotes 3cc0957) Update a new skill description Fix bugs

+++ app/src/main/java/com/github/malitsplus/shizurunotes/data/action/RevivalAction.java
@@ -40,7 +40,7 @@ public class RevivalAction extends ActionParameter {
         switch (revivalType){
             case normal:
                 return I18N.getString(R.string.Revive_s1_with_d2_HP,
-                        targetParameter.buildTargetClause(), Math.ceil(actionValue2.value * 100));
+                        targetParameter.buildTargetClause(), (int) Math.ceil(actionValue2.value * 100));
             default:
                 return super.localizedDetail(level, property);
         }
